requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.5
essentia-tensorflow==2.1b6.dev1389
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.1
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
protobuf==4.25.8
Pygments==2.19.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
six==1.17.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.2
zipp==3.23.0
